refactor: de-meteorize sendWelcomeEmail util
